Add interface and future implementation
